more conv/autoencoder experiments
thought about this space-to-depth paper (https://github.com/LabSAINT/SPD-Conv) throughout the family
vacation and eventually tried it out. well, it might be an advantage for small-scale image classification
but it does not seem to change much for autoencoder quality. but what do i know?

also tried the dalle-2 autoencoder which is of course much better than anything i came up with,
but it's also ten times slower.

currently thinking about tiled game graphics generation via the decoder part. using CLIP to guide
the code into a direction, one can create A LOT of new graphics, each only 128 numbers for storage
and you can randomize or morph the code to create slight (or strong) variations of the graphics...