tried to implement a renderer for s3m tracker modules
would be fantastic to have such a thing, but current version is too slow to be used
about 40x real-time :-(

it's not even the number crunching. simply iterating through all channels and
increasing each channel's `samplepos` is slower than realtime.

i'd like to refactor existing code to render audio buffers via numpy or torch
(that was the original plan) but that's a big task..