the "conditional denoiser" sessions
tried adding a condition in form of class logits to my simple ResConv
which actually works quite well in the case of MNIST denoising.
However, actual image generation is complicated.

then expanded on the huggingface tutorial for diffusion training which
works much better in generation :thumbsup:
However, how to actually control these things. It's not like massaging
some pixels via CLIP targets. it's quite complicated to add the right
amount of noise to an existing image to control the diffuser. it basically
just creates whatever it wants (from the requested class)