blog: fix the json url
ah well, this is some slight problem i might fix in the future..
currently, for development i'm just running `python -m http.server` inside the docs/ directory
but on ghpages everything is prefixed with the /nn-experiments path.
it works for markdown links (they are all converted to relative paths via ../..)
but, of course, not for inline urls in javascript files. the js file would need to fetch
the url from the rendered html, preferably from a data-url attribute but these attributes
are not handled by the html renderer..